[REFACTOR] 주문 페이지 리팩토링
