+++ src/constants/mode.js
@@ -7,3 +7,8 @@ export const FAQ_MODE = {
   PRODUCT_WARRANTY: 'PRODUCT_WARRANTY',
   ETC: 'ETC',
 };
+
+export const ORDER_MODE = {
+  CART: 'cart',
+  BUY: 'buy',
+};
