+++ src/constants/path.js
@@ -1,8 +1,21 @@
+export const MAIN = `/`;
+export const MY_INFO = `/myInfo`;
+export const LOGIN = `/login`;
+export const SIGN_UP = `/signUp`;
+export const REDIRECT = `/redirect`;
+export const CART = `/cart`;
+export const CHOOSE = `/choose`;
+export const ORDER = `/order`;
+export const PAYMENT = `/payment`;
+export const CHANGE_PASSWORD = `/auth/password`;
+export const INQUIRY = `/inquiry`;
+export const DETAIL = `/detail/:orderId`;
+export const GUEST_DETAIL = `/product/guest/:orderId`;
+export const EDITOR = `/editor`;
 export const SUPPORT = `/support`;
 export const QUESTION = '/question';
 export const QUESTION_DETAIL = `/question/:id`;
 export const QUESTION_EDITOR = `/question/editor`;
 export const FAQ = `/faq`;
+export const BACK = `/`;
 export const PRODUCT = `/products`;
-export const BACK = '/';
-export const CART = '/cart';
